requirements.txt
boto3>=1.34.0
pdfplumber>=0.10.0
chromadb>=0.5.0
langchain>=0.2.0
langchain-aws>=0.1.0
langchain-community>=0.2.0
langchain-chroma>=0.1.0
streamlit>=1.35.0
pandas>=2.0.0
faker>=25.0.0
reportlab>=4.0.0
python-dotenv>=1.0.0
clickhouse-connect>=0.7.0
